backend/requirements.txt
django==2.0.5
djangorestframework==3.8.2
django-cors-headers==2.2.0
django-autofixture==0.12.1
